changes to download_files to detect existing zip or csv

+++ retinaburner/usaspending/management/commands/download_files.py
@@ -11,10 +11,14 @@ class Command(BaseCommand):
         csv_path = settings.CSV_PATH
         for line in open(download_file).xreadlines():
             line = line.strip()
-            print "Reading file " + line
+            print "Reading line " + line
             outfile_name = csv_path + line.split('/')[-1]
-            print "saving to " + outfile_name
-            urllib.urlretrieve(line, outfile_name)
+            print csv_path + 'datafeeds/' + line.split('/')[-1].replace('.zip', '')
+            if os.path.exists(outfile_name) or os.path.exists(csv_path + 'datafeeds/' + line.split('/')[-1].replace('.zip', '')):
+                print "File exists, skipping"
+            else:
+                print "saving to " + outfile_name
+                urllib.urlretrieve(line, outfile_name)
 
         for f in os.listdir(csv_path):
             if f[-3:] == 'zip':
